Remove AGG.
It's no longer needed because i now use my own port of Gnash
that provides its own copy of AGG and it was the only piece of software
using it.
This is also done for performance reasons as that means the whole AGG
library can easily be PGO'd, something that would be a lot less easy
if part of the OS.

To ease maintaince on the Funkey team and given that it was only useful for Gnash,
it should be removed.